Move `/help` to `/help/aquifer-cms/documents`
